WS-74: types extracted
